Do not double convert yt videos

+++ main.py
@@ -208,6 +208,7 @@ async def send_converted_video(context: ContextTypes.DEFAULT_TYPE):
     chat_id = job.chat_id
     data = job.data["data"]
     is_file_name = job.data["is_file_name"]
+    need_conversion = job.data.get("is_need_conversion") or True
     caption = job.data.get("caption")
     if is_file_name:
         original = data
@@ -217,7 +218,10 @@ async def send_converted_video(context: ContextTypes.DEFAULT_TYPE):
         raise ProcessException(f"Can't download video from {data}")
     is_nsfw = any(flag in data for flag in NSFW_FLAGS)
     try:
-        converted = await convert2MP4(original)
+        if need_conversion:
+            converted = await convert2MP4(original)
+        else:
+            converted = original
         file_size_megabytes = os.path.getsize(converted) / (1024 * 1024)
         if file_size_megabytes > 50:
             raise ProcessException(
@@ -238,9 +242,9 @@ async def send_converted_video(context: ContextTypes.DEFAULT_TYPE):
     except Exception:
         raise
     finally:
-        if original:
+        if original and original != converted:
             remove_file(original)
-        if converted:
+        if converted and original != converted:
             remove_file(converted)
 
 
@@ -432,7 +436,7 @@ async def send_instagram_video(context: ContextTypes.DEFAULT_TYPE):
         send_converted_video,
         1,
         chat_id=chat_id,
-        data=dict(data=reel_filename, is_file_name=True, caption=link),
+        data=dict(data=reel_filename, is_file_name=True, caption=link, is_need_conversion=True),
     )
 
 
@@ -475,7 +479,7 @@ async def send_youtube_video(context: ContextTypes.DEFAULT_TYPE):
         send_converted_video,
         1,
         chat_id=chat_id,
-        data=dict(data=video_filename, is_file_name=True, caption=link),
+        data=dict(data=video_filename, is_file_name=True, caption=link, is_need_conversion=False),
     )
 
 
@@ -490,7 +494,7 @@ async def send_tiktok_video(context: ContextTypes.DEFAULT_TYPE):
         send_converted_video,
         1,
         chat_id=chat_id,
-        data=dict(data=video_filename, is_file_name=True, caption=link),
+        data=dict(data=video_filename, is_file_name=True, caption=link, is_need_conversion=True),
     )
